tweaking on road trip
