+++ recordings/models/__init__.py
@@ -1,0 +1,15 @@
+from recordings.models.course import (
+    Course, CourseResource, CourseSession, SessionReference
+)
+from recordings.models.models import (
+    ClassRecording, CodeWarsProblem
+)
+
+
+__all__ = (ClassRecording.__name__,
+           CodeWarsProblem.__name__,
+           Course.__name__,
+           CourseSession.__name__,
+           CourseResource.__name__,
+           SessionReference.__name__,
+)
