+++ recordings/models/_base.py
@@ -1,0 +1,13 @@
+from django.db import models
+
+
+class CourseResource(models.Model):
+    name = models.CharField(max_length=250, null=True, blank=True)
+    description = models.TextField(null=True, blank=True)
+    created = models.DateTimeField(auto_now_add=True)
+    modified = models.DateTimeField(auto_now=True)
+
+    class Meta:
+        abstract = True
+
+
